refactor(tagbookmark): TagBookmarkController에 역할 분담
